requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
python-dotenv==0.19.0